opnode: Restructure Block References
This flattens the Block ID fields and adds a time field to
the Block References. The time field is important for sequencing.
It needs to be in the BlockRef instead of the BlockID because
it is hard to reliably fill out the timestamp field for
Parent and L1Origin block ids.

This approach also removes the ParentNumber because it can
be easily calculated if needed, but was not being used.